LPS-53223 New pattern
