+++ modules/apps/document-library/document-library-ratings-definition/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Liferay Document Library Ratings Definition
-Bundle-SymbolicName: com.liferay.documentlibrary.ratings.definition
+Bundle-SymbolicName: com.liferay.document.library.ratings.definition
 Bundle-Version: 1.0.0
 Include-Resource: classes
