+++ modules/apps/message-boards/message-boards-ratings-definition/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Liferay Message Boards Ratings Definition
-Bundle-SymbolicName: com.liferay.messageboards.ratings.definition
+Bundle-SymbolicName: com.liferay.message.boards.ratings.definition
 Bundle-Version: 1.0.0
 Include-Resource: classes
